requirements.txt
google-genai>=1.0.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
